task: removed IPrimitiveSerializer interface
